+++ src/run/ui_main.py
@@ -5,7 +5,6 @@ sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
 from PyQt5.QtWidgets import QApplication
 from loguru import logger
-from ui_event.base_event import exit_save_data, kill_other_close
 from log.log_record import log_record, check_version
 from ui_event.pyqt_main_ui import UIMainWindows
 from utils.sytem_monitor import sys_mon
@@ -20,12 +19,8 @@ def ui_paint():
     """
     app = QApplication(sys.argv)
 
-    # app.setQuitOnLastWindowClosed(True)
-
     # 单次点击close不退出
     app.setQuitOnLastWindowClosed(False)
-
-    # app.lastWindowClosed.connect(on_last_window_closed)
 
     w = UIMainWindows()
     w.show()