+++ src/ui_event/img_analyze_dialog.py
@@ -40,7 +40,8 @@ class ImgAnalyzeHistogram(QDialog):
         self.init_ui("Image Category Analyze")
         self.updateChart()
 
-    def init_ui(self, window_title):
+    @logger.catch
+    def init_ui(self, window_title, _=None):
         """
 
         :return: 
@@ -67,7 +68,8 @@ class ImgAnalyzeHistogram(QDialog):
         self.layout.addLayout(self.h_layout)
         self.layout.addWidget(self.next_button)
 
-    def parse_img_data(self):
+    @logger.catch
+    def parse_img_data(self, _=None):
         """
 
         :return:
@@ -76,7 +78,8 @@ class ImgAnalyzeHistogram(QDialog):
         img_name_list, img_count_list = img_analyze_data_output_new()
         return img_count_list, img_name_list
 
-    def create_chart(self):
+    @logger.catch
+    def create_chart(self, _=None):
         """
         创建并返回一个带有数据标签的柱状图
         :return: QChart 对象
@@ -95,7 +98,8 @@ class ImgAnalyzeHistogram(QDialog):
 
         return self.chart
 
-    def create_pie_chart(self):
+    @logger.catch
+    def create_pie_chart(self, _=None):
         self.series_pie = QPieSeries()
 
         self.pie_chart = QChart()
@@ -105,7 +109,8 @@ class ImgAnalyzeHistogram(QDialog):
 
         return chart_view
 
-    def showNextGroup(self):
+    @logger.catch
+    def showNextGroup(self, _=None):
 
         try:
             self.current_group = (self.current_group + 1) % (len(self.error_counts) // self.group_size)
@@ -114,7 +119,8 @@ class ImgAnalyzeHistogram(QDialog):
 
         self.updateChart()
 
-    def updateChart(self):
+    @logger.catch
+    def updateChart(self, _=None):
         """
 
         :return:
@@ -168,9 +174,11 @@ class ImgAnalyzeHistogram(QDialog):
         self.pie_chart.addSeries(self.series_pie)
         self.pie_chart.setTitle("pie chart")
 
-    def closeEvent(self, event):
+    @logger.catch
+    def closeEvent(self, event, _=None):
         """
         对话框关闭
+        :param _:
         :param event:
         :return:
         """